fix: Fixed an API sqlite3 error
Fixed a programming error that would create exceptions whenever the ID
counter would enter double digits.

Also updated the code for the POST response to return the product with
the highest ID value in the table, rather than the name and description
like it was previously, which was less reliable.