fix: Fixed programming error with API requests to DB
All of the calls to the database that used parameter substitution for
only one value were improperly formatted and would create errors in some cases.

Also adds a fix where the function that handles the POST request to
/api/products will now form the response data for the added product by
selecting the product with the highest ID number and not the same name
and description. This is more reliable and prevents certain errors from
occurring when multiple entries with identical data are added, usually
as a result from testing.